CovidApp/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pytz==2021.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
translator==0.0.9
transliterate==1.10.2
tzdata==2021.5
urllib3==1.26.8
